requirements.txt
yt-dlp>=2024.0.0
openai-whisper
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
requests
moviepy==1.0.3
deep-translator
beautifulsoup4
